# hw2_openAPI_CRUD/openapi/marketplace.yaml
openapi: "3.1.0"
info:
  title: Marketplace API
  version: "1.0.0"
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
    ErrorCode:
      type: string
      enum:
        - PRODUCT_NOT_FOUND
        - PRODUCT_INACTIVE
        - ORDER_NOT_FOUND
        - ORDER_LIMIT_EXCEEDED
        - ORDER_HAS_ACTIVE
        - INVALID_STATE_TRANSITION
        - INSUFFICIENT_STOCK
        - PROMO_CODE_INVALID
        - PROMO_CODE_MIN_AMOUNT
        - ORDER_OWNERSHIP_VIOLATION
        - VALIDATION_ERROR
        - TOKEN_EXPIRED
        - TOKEN_INVALID
        - REFRESH_TOKEN_INVALID
        - ACCESS_DENIED
        - CONFLICT
    ErrorResponse:
      type: object
      required:
        - error_code
        - message
      properties:
        error_code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Error message
        details:
          type: object
          nullable: true
          description: Error details
    ProductStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
        - ARCHIVED
    ProductCreate:
      type: object
      required:
        - name
        - price
        - stock
        - category
        - status
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          description: Product name
        description:
          type: string
          maxLength: 4000
          nullable: true
          description: Product description
        price:
          type: number
          format: decimal
          minimum: 0.01
          description: Unit price
          example: 99.99
        stock:
          type: integer
          minimum: 0
          description: Stock quantity
          example: 100
        category:
          type: string
          minLength: 1
          maxLength: 100
          description: Product category
          example: electronics
        status:
          $ref: '#/components/schemas/ProductStatus'
    ProductUpdate:
      type: object
      description: Partial update
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          nullable: true
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
          nullable: true
        stock:
          type: integer
          minimum: 0
          nullable: true
        category:
          type: string
          minLength: 1
          maxLength: 100
          nullable: true
        status:
          $ref: '#/components/schemas/ProductStatus'
          nullable: true
    ProductResponse:
      type: object
      required:
        - id
        - name
        - price
        - stock
        - category
        - status
        - seller_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Product identifier
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: decimal
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
          description: Seller user ID
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
    ProductListResponse:
      type: object
      required:
        - items
        - total_elements
        - page
        - size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        total_elements:
          type: integer
          minimum: 0
          description: Total matching products
        page:
          type: integer
          minimum: 0
          description: Current page (0-based)
        size:
          type: integer
          minimum: 1
          description: Page size
...
    UserRole:
      type: string
      enum:
        - USER
        - SELLER
        - ADMIN
    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
          description: Unique username
        email:
          type: string
          format: email
          description: Email address
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 128
          description: Password
        role:
          type: string
          enum:
            - USER
            - SELLER
          description: User role
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          description: Refresh token
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        token_type:
          type: string
          default: bearer
    AccessTokenResponse:
      type: object
      required:
        - access_token
        - token_type
      properties:
        access_token:
          type: string
          description: JWT access token
        token_type:
          type: string
          default: bearer
    UserResponse:
      type: object
      required:
        - id
        - username
        - email
        - role
        - created_at
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Invalid or expired token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            token_expired:
              value:
                error_code: TOKEN_EXPIRED
                message: Access token has expired
                details: null
            token_invalid:
              value:
                error_code: TOKEN_INVALID
                message: Invalid access token
                details: null
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: ACCESS_DENIED
            message: "Role USER not allowed. Required: SELLER, ADMIN"
            details: null
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error_code: VALIDATION_ERROR
            message: Request validation failed
            details:
              fields:
                - field: price
                  message: Input should be greater than or equal to 0.01
                  type: greater_than_equal
paths:
  /auth/register:
    post:
      operationId: register
      summary: Register a new user
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "409":
          description: Username or email conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      operationId: login
      summary: Authenticate and obtain tokens
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: Tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        "200":
          description: New access token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTokenResponse'
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: REFRESH_TOKEN_INVALID
                message: Refresh token is invalid or expired
                details: null
  /products:
    post:
      operationId: createProduct
      summary: Create a new product
      description: "SELLER (own), ADMIN"
      tags:
        - products
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'

    get:
      operationId: listProducts
      summary: List products with pagination and filtering
      description: Excludes ARCHIVED by default
      tags:
        - products
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page number (0-based)
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Page size
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
          description: Status filter
        - name: category
          in: query
          schema:
            type: string
          description: Category filter (exact)
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
  /products/{id}:
    get:
      operationId: getProduct
      summary: Get product by ID
      description: "All roles"
      tags:
        - products
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error_code: PRODUCT_NOT_FOUND
                message: Product not found
                details: null

    put:
      operationId: updateProduct
      summary: Update product
      description: "SELLER (own), ADMIN"
      tags:
        - products
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        "400":
          $ref: '#/components/responses/ValidationError'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: deleteProduct
      summary: Soft-delete product (sets status to ARCHIVED)
      description: "SELLER (own), ADMIN"
      tags:
        - products
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Product archived
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...